third_party/blink/web_tests/external/wpt/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-remove-head-expected.txt

This is a testharness.js-based test.
FAIL Meta color-scheme applies. assert_equals: Root element's color-scheme should be 'normal' expected "normal" but got "dark"
PASS Initial value after removing head including meta color-scheme.
Harness: the test ran to completion.
